requirements.txt
# Coex Compiler Dependencies

# Parser generator runtime
antlr4-python3-runtime>=4.13.0

# LLVM bindings for code generation
llvmlite>=0.41.0

# Testing framework
pytest>=7.0.0

# Optional: improved test output
pytest-sugar>=0.9.0
